requirements.txt
bcrypt
cryptography
twilio
firebase-admin
opencv-python
numpy
mtcnn
web3
ttkbootstrap
Pillow
requests
matplotlib